Refactor: Add Tags as component and update related css
